app/pyproject.toml
[tool.poetry]
name = "task-service"
version = "0.1.0"
description = ""
authors = ["sxkxixx <sasha.kornilov.1212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.1.1"}
python-multipart = "^0.0.6"
aioboto3 = "^11.2.0"
pytest-asyncio = "0.21.1"
asgi-lifespan = "^2.1.0"
httpx = "^0.24.1"
aioredis = "^2.0.1"
gunicorn = "^21.2.0"
sse-starlette = "^1.6.5"
celery = "^5.3.4"
redis = "^5.0.1"
pytest = "7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
